feat: add `handleChord` function and update doc comments
